fix reintroduced race(?) condition
Calling the emit functions within the foreach loop consequently reintroduced the error in exactly the same place it was before, albeit now with the previous changes from #205fc55 and #106b4f3 the fix is easier. This code (like previously) works fine only when focusing the monitor with the greatest id, i.e., last in the array, otherwise the emit is too early and sends the 'changed' signal with only a partially construction monitors array. Moving the emits to after the foreach should rectify this, albeit I don't currently have the means to test it myself.